Fixing hang tests for CI/CD.
